feat: added a checkbox to enable visuals on verify
A checkbox to decide whether the visuals are on or not while running the
bot to verify the level has been added. Now the verify proccess shows
the number of attempts and the max percentage reached instead of the
percentage of the last attempt. Now the windows close themselves after
the accept button is pressed.